% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hwsd2}
\alias{hwsd2}
\title{HWSD v2.0 database}
\format{
data.frame
}
\usage{
hwsd2
}
\description{
Database holding the full HWSD v2.0 database layer information
for the main soil type specified. For the fields included (i.e.
the column names I refer to the FAO documentation).
}
\keyword{datasets}
