% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hwsd_meta_data}
\alias{hwsd_meta_data}
\title{HWSD meta-data}
\format{
data.frame
\describe{
  \item{parameter}{tile ID number}
  \item{subset}{minimum longitude}
  \item{description}{maximum longitude}
  \item{units}{minimum latitude}
}
}
\usage{
hwsd_meta_data
}
\description{
Data frame with meta-data on the ORNL DAAC parameters one can query
using the THREDDS server. In addition a brief description of the
various data products and their units is provided.
}
\keyword{datasets}
