% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get.R
\name{ws_get}
\alias{ws_get}
\title{Basic HWSD download function}
\usage{
ws_get(location, param, path, internal = TRUE)
}
\arguments{
\item{location}{file with several site locations and coordinates
in a comma delimited format: site, latitude, longitude}

\item{param}{which soil parameter to use}

\item{path}{default is tempdir()}

\item{internal}{return an internal raster or just retain values in the path}
}
\value{
HWSD data
}
\description{
Downloads HWSD data, wrapped by ws_subset() for convenient use
}
