% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_subset.R
\name{ws_subset}
\alias{ws_subset}
\title{Function to geographically subset 'Daymet' regions exceeding tile limits}
\usage{
ws_subset(
  site = "HWSD",
  location = c(34, -81, 32, -80),
  param = "ALL",
  path = tempdir(),
  internal = TRUE
)
}
\arguments{
\item{site}{sitename for the extracted location}

\item{location}{location of a bounding box c(lat, lon, lat, lon) defined
by a top left and bottom-right coordinates, a single location (lat, lon)
or a data frame with various locations listed (site, lat, lon)}

\item{param}{soil parameters to provide, the default setting is ALL, this 
will download all available soil parameters.Check
https://daac.ornl.gov/SOILS/guides/HWSD.html for parameter descriptions.}

\item{path}{path where to download the data to (only applicable to
spatial data)}

\item{internal}{do not store the data on disk}
}
\value{
netCDF data, or a data frame with HWSD soil information
}
\description{
Function to geographically subset 'Daymet' regions exceeding tile limits
}
\examples{

\dontrun{
print("test")
}
}
